+++ backend/src/routes/friendSuggestion.ts
@@ -2,38 +2,11 @@ import { Router } from "express";
 import prisma from "../prisma";
 import requireAuth from "../middleware/requireAuth";
 import { recipeMap } from "../utils/recipeMap";
+import {
+  top5Users
+} from "../utils/similarity"
 
 const router = Router();
-
-//calculates recipe similarity using weighted jaccard similarity
-//weight: 1 (if saved), 2 (if liked), 3 (if saved and liked)
-function recipeSimilar(
-  mapA: Map<string, number>,
-  mapB: Map<string, number>
-): number {
-  const allRecipeIds = new Set([...mapA.keys(), ...mapB.keys()]);
-  let intersection = 0;
-  let union = 0;
-  for (const id of allRecipeIds) {
-    const a = mapA.get(id) || 0;
-    const b = mapB.get(id) || 0;
-    intersection += Math.min(a, b); // both users saved recipes
-    union += Math.max(a, b); //unique recipes
-  }
-  return union === 0 ? 0 : intersection / union;
-}
-
-//ingredient overlap similarity using jaccard
-function ingredientOverlap(
-  ingredientsA: Set<string>,
-  ingredientsB: Set<string>
-): number {
-  const intersection = new Set(
-    [...ingredientsA].filter((ing) => ingredientsB.has(ing))
-  );
-  const union = new Set([...ingredientsA, ...ingredientsB]);
-  return union.size === 0 ? 0 : intersection.size / union.size;
-}
 
 router.get("/friends/suggestions", requireAuth, async (req, res) => {
   const username = req.session.user?.username;
@@ -53,37 +26,20 @@ router.get("/friends/suggestions", requireAuth, async (req, res) => {
       return;
     }
 
-    const userIngredients = new Set<string>();
-    const userMap = recipeMap(currUser.saved, currUser.liked, userIngredients);
-
     const otherUsers = await prisma.user.findMany({
       where: { id: { not: currUser.id } },
       include: { saved: true, liked: true },
     });
-    const suggestions = otherUsers
-      .map((user) => {
-        const otherIngredients = new Set<string>();
-        const otherMap = recipeMap(user.saved, user.liked, otherIngredients);
 
-        //calculating similarity score for both recipe and ingredient based
-        const recipeScore = recipeSimilar(userMap, otherMap);
-        const ingredientScore = ingredientOverlap(
-          userIngredients,
-          otherIngredients
-        );
-        const finalScore = 0.7 * recipeScore + 0.3 * ingredientScore;
+    const topUsers = top5Users(currUser, otherUsers);
 
-        return {
-          id: user.id,
-          username: user.username,
-          similarity: parseFloat(finalScore.toFixed(2)),
-          recipeScore: parseFloat(recipeScore.toFixed(2)),
-          ingredientScore: parseFloat(ingredientScore.toFixed(2)),
-        };
-      })
-      .filter((s) => s.similarity > 0.1) //filter out users who have little similarity
-      .sort((a, b) => b.similarity - a.similarity) //sort from highest similarity
-      .slice(0, 5); //top 5 users with the highest similarity
+    const suggestions = topUsers.map(({user, finalScore, recipeScore, ingredientScore}) => ({
+      id: user.id,
+      username: user.username,
+      similarity: parseFloat(finalScore.toFixed(2)),
+      recipeScore: parseFloat(recipeScore.toFixed(2)),
+      ingredientScore: parseFloat(ingredientScore.toFixed(2)),
+    }))
 
     res.json(suggestions);
   } catch (err) {
